+++ builder/builders/compilers/typescript.ts
@@ -19,23 +19,14 @@ export default class TypescriptCompiler extends Compiler {
     const exportedPath = exportPath;
     const result = await this.bundler.bundle();
 
-    return {
+    return this.postCompile({
       output: result.output,
       file: this.file,
       path: exportedPath,
       type: 'js' as ExportType,
       affectedFiles: []
-    };
-  }
-
-  public async getContextFiles(
-    exportPath: string,
-    context: IBuilderContext
-  ): Promise<string[]> {
-    const bundler = new TypescriptBundler(this.file);
-    const result = await bundler.bundle();
-    return [
+    }, [
       ...result.modules.filter((v: any) => !v.node_module).map((v) => v.file as { file: string, nodeModule: boolean })
-    ].filter(v => !!v && !v.nodeModule).map(v => v.file);
+    ].filter(v => !!v && !v.nodeModule).map(v => v.file));
   }
 }