+++ cli/dev-server/index.ts
@@ -1,20 +1,23 @@
-import chalk from 'chalk';
-import fs from 'fs';
-import path from 'path';
-import { from, Subscription, timer } from 'rxjs';
-import { filter, mergeMap, tap } from 'rxjs/operators';
+import chalk from "chalk";
+import fs from "fs";
+import path from "path";
+import { from, Subscription, timer } from "rxjs";
+import { filter, mergeMap, tap } from "rxjs/operators";
 
-import { Builder } from '../../builder';
-import { ExportType } from '../../builder/builders/utils';
-import { IBuilderOptions } from '../../builder/definitions/builder-options';
-import { IMainCommanderOptions, resolveFilePath, resolveFilePathOnBase } from '../utils';
-import { CompilationMap, CompilationStatus } from './compilation-map';
-import { WebServer } from './web-server';
+import { Builder } from "../../builder";
+import { ExportType } from "../../builder/builders/utils";
+import { IBuilderOptions } from "../../builder/definitions/builder-options";
+import {
+  IMainCommanderOptions,
+  resolveFilePath,
+  resolveFilePathOnBase,
+} from "../utils";
+import { CompilationMap, CompilationStatus } from "./compilation-map";
+import { WebServer } from "./web-server";
 
 const log = console.log;
 
 export class DevServer {
-
   private readonly builder: Builder;
   public readonly webSever?: WebServer;
 
@@ -29,10 +32,13 @@ export class DevServer {
     public readonly options: IBuilderOptions,
     public readonly devServerOptions = {
       port: 5678,
-      socketPort: 5679
+      socketPort: 5679,
     }
   ) {
-    this.builder = new Builder(options, path.dirname(resolveFilePathOnBase(commander.config)));
+    this.builder = new Builder(
+      options,
+      path.dirname(resolveFilePathOnBase(commander.config))
+    );
 
     this.files = new CompilationMap(options);
 
@@ -40,22 +46,38 @@ export class DevServer {
 
     this.start();
   }
+
   public get Files() {
-    return this.files.asObservable(); 
+    return this.files.asObservable();
   }
-  
+
+  private getRelativeSourceFileName(file: string) {
+    return path.relative(this.builder.basePath, file);
+  }
+
   private async compileFile(file: string) {
     const exportPath = this.builder.getExportPathOfFile(file);
-    if (!exportPath) { return; }
+    if (!exportPath) {
+      return;
+    }
     const eFile = path.relative(this.builder.options.output, exportPath);
 
     this.files.changeFile(eFile);
-    
-    const result = (await this.builder.compile(
-      async (localFile: { path: string; type: ExportType; affectedFiles: string[]; }) =>
-        new Promise<void>(resolve => { console.log('compiled', file); resolve() }),
-      [file]
-    ))[0];
+
+    const result = (
+      await this.builder.compile(
+        async (localFile: {
+          path: string;
+          type: ExportType;
+          affectedFiles: string[];
+        }) =>
+          new Promise<void>((resolve) => {
+            console.log(chalk.green("compiled", this.getRelativeSourceFileName(file), '->', this.getRelativeSourceFileName(exportPath)));
+            resolve();
+          }),
+        [file]
+      )
+    )[0];
 
     const fileOutput = this.files.getValue().get(eFile);
     if (fileOutput) {
@@ -63,68 +85,63 @@ export class DevServer {
         ...fileOutput,
         output: this.webSever?.processFile(result).output || result.output,
         compilationStatus: CompilationStatus.Compiled,
-        changeAmount: fileOutput.changeAmount
+        changeAmount: fileOutput.changeAmount,
       });
     } else {
       this.files.patchFile(eFile, {
         ...result,
         output: this.webSever?.processFile(result).output || result.output,
         compilationStatus: CompilationStatus.Compiled,
-        changeAmount: 1
+        changeAmount: 1,
       });
     }
   }
 
-  // TODO: Somehow make the builder push changes rather than checking for them.
-  public getWatchingFilesObservable() {
-    let files: { source: string; files: string[]; }[] = [];
-    return timer(0, 2000).pipe(
-      mergeMap(() => from(this.builder.getContextFiles())),
-      filter(v => JSON.stringify(v) !== JSON.stringify(files)),
-      tap(v => files = v)
-    )
-  }
-
-  public async getWatchingFiles() {
-    return await this.builder.getContextFiles()
-  }
-
   private async start() {
-    this.watcherSub = this.getWatchingFilesObservable().subscribe(async contextFiles =>{
-      
-      // const contextFiles = this.getWatchingFiles();
-        
-      log('Watching files');
-      contextFiles.map(v => v.files.map(file => chalk.underline.blue(file)).join('\n  ')).forEach(v => log('  ' + v));
-      log('');
+    this.watcherSub = this.builder.ContextFiles.subscribe(
+      async (contextFiles) => {
+        log("Watching files");
+        contextFiles
+          .map(({ files, source }) =>
+            chalk.underline.blue(
+              `${this.getRelativeSourceFileName(source)}${
+                files.length > 0
+                  ? chalk.cyan` \n    watched through context (require or import):\n     - ${files
+                      .map((file) => this.getRelativeSourceFileName(file))
+                      .join("\n     - ")}`
+                  : ""
+              }`
+            )
+          )
+          .forEach((v) => log("  " + v));
+        log("");
 
-      this.watchers.forEach(watcher => {
-        watcher.close();
-      })
-  
-      await contextFiles.forEach(async context => {
-        [ ...context.files ].forEach(file => {
-          try {
-            this.watchers.push(fs.watchFile(file, (curr: any, prev: any) => {
-              this.compileFile(context.source);
-            }) as any);
-          } catch (e) {
-            console.log('FILE', file);
-            console.log(e)
-          }
+        this.watchers.forEach((watcher) => {
+          watcher.close();
         });
-        await this.compileFile(context.source);
-      });
 
-    })
-    
-
+        await contextFiles.forEach(async (context) => {
+          [...context.files].forEach((file) => {
+            console.log('watching file', file)
+            try {
+              this.watchers.push(
+                fs.watch(file, (curr: any, prev: any) => {
+                  this.compileFile(context.source);
+                })
+              );
+            } catch (e) {
+              console.log("FILE", file);
+              console.log(e);
+            }
+          });
+          await this.compileFile(context.source);
+        });
+      }
+    );
   }
-  
 
   public async destroy() {
     this.watcherSub?.unsubscribe();
     return this.webSever!.destroy();
   }
-
 }