+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+### Fixed
+- Memory leak with files being constantly recompiled when watching or running a dev-server.
+- Cleaned up debug output of console on dev-server.
+- Fixed output paths of watch and build commands.
+
+### Changed
+- Context file detection for watch and dev-server are now self-maintaining within the builders.
+- Context file detection for watch and dev-server is now an RXJS observable.
+
 ## 0.7.4 - 8. November 2020
 ### Added
 - Added root as option to optionally define the root of your project.